File Writter

+++ utils.py
@@ -61,7 +61,11 @@ def creative_name_search(keyword,target):
     
 
 def  write_excel_data(response,file_path,traffic):
-    df2 = pd.DataFrame(response)
-    with pd.ExcelWriter(file_path) as writer:  
-        df2.to_excel(writer, sheet_name=str(traffic['Operative_ID']))
-        return True
+    if(os.path.exists(file_path) == True):
+        df2 = pd.DataFrame(response)
+        with pd.ExcelWriter(file_path, engine='openpyxl', mode='a') as writer:  
+            df2.to_excel(writer, sheet_name=str(traffic['Operative_ID']))
+    else:
+        df2 = pd.DataFrame(response)
+        with pd.ExcelWriter(file_path) as writer:  
+            df2.to_excel(writer, sheet_name=str(traffic['Operative_ID']))
